requirements.txt
aws-cdk-lib==2.50.0 #92.0
boto3==1.24.64
constructs==10.0.37
dynamodb-json==1.3
envs==1.4
install==1.3.5
